fixed #63
